// src/lib/types/character.ts
export type EquipmentSlot = 'head' | 'torso' | 'legs' | 'hands' | 'feet' | 'ring' | 'amulet' | 'weapon';
// ...
export type ItemType = 'weapon' | 'armor' | 'potion' | 'misc' | 'gold' | 'spell_tome' | 'key_item' | 'learning_book' | 'food';

export interface CharacterInventoryItem {
    id: string; // The ID of the base item from the 'items' collection
    name: string;
    weight: number;
    quantity: number;
    type: ItemType;
    rarity?: Rarity;
    equipmentSlot?: EquipmentSlot;
    damage?: number;
    damageDice?: string; // e.g., "1d8"
    armor?: number;
    effect?: {
        id?: string; // Unique ID for the effect, used to check if it's already active
        type: 'heal' | 'buff';
        stat: string;
        amount: number;
        duration?: number; // in ms
        description?: string;
        icon?: string;
    };
    spellId?: string;
    learningEffect?: {
        id: string;
        name: string;
        description: string;
        duration: number;
        icon: string;
    };
}